pyproject.toml
[tool.poetry]
name = "auth0-token"
version = "0.1.2"
description = "A python library to get an auth0 access token"
authors = ["KaaS Team"]

[tool.poetry.dependencies]
python = "^3.7"
auth0-python="^3.19.0"
cachetools="^5.0.0"
cryptography = "36.0.1"

[tool.poetry.dev-dependencies]
pytest = "^7.3.1"
pytest-cov = "^2.10.1"
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
